# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: specs-generator
# during dev, should point to your local machine
host: localhost:8000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /generate:
    # binds a127 app logic to a route
    x-swagger-router-controller: generators
    post:
      description: Generate 'API' specs from CURL format.
      # used as the method name of the controller
      operationId: generateFromCURL
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - name: body
          in: body
          required: true
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GenerateRequest"
      responses:
        "201":
          description: Created
        # responses may fall through to errors
        default:
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
  /cors:
    # binds a127 app logic to a route
    x-swagger-router-controller: generators
    get:
      description: Get 'API' resource from browser request (CORS passthrough).
      # used as the method name of the controller
      operationId: getCORS
      parameters:
        - name: url
          in: query
          required: true
          type: string
      responses:
        "200":
          description: Success
        # responses may fall through to errors
        default:
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GenerateRequest:
    required:
      - type
    properties:
      type:
        type: string
      data:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string